% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableVarieties.R
\name{listAvailableVarieties}
\alias{listAvailableVarieties}
\title{List Available Calibrated Varieties}
\usage{
listAvailableVarieties(user_dir = NULL)
}
\arguments{
\item{user_dir}{Optional. Path to the folder containing user-calibrated parameter files.}
}
\value{
A character vector of available calibrated variety filenames.
}
\description{
Lists the CSV files containing calibrated parameters included in the package
and any additional user-calibrated files in a given directory.
}
