% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\docType{data}
\name{phenomenalsParameters}
\alias{phenomenalsParameters}
\title{Default phenology parameters}
\format{
A named list of species → class → parameter → list(min, max, value, calibration).
}
\usage{
phenomenalsParameters
}
\description{
A nested list of default parameters used in the Phenomenals model.
}
\examples{
phenomenalsParameters[["grapevine"]][["phenomenalsParameters"]][["TminPhenology"]]
}
\keyword{datasets}
